restructured GameBoard logic, added custom gson deserializer for GameStarted

+++ RoboRally/src/main/java/utilities/messages/PlayerStatus.java
@@ -2,17 +2,17 @@ package utilities.messages;
 
 public class PlayerStatus extends MessageBody {
 
-    private final double playerID;
+    private final int playerID;
     private final Boolean ready;
 
-    public PlayerStatus(double playerID, Boolean ready) {
+    public PlayerStatus(int playerID, Boolean ready) {
 
         this.playerID = playerID;
         this.ready = ready;
 
     }
 
-    public double getPlayerID() {
+    public int getPlayerID() {
         return playerID;
     }
